+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * Added `__delitem__` to Meta and `drop` to MetaHeader and MetaLabels classes.
   * Modified Meta to allow MetaHeader attribute access directly from Meta.
   * Expanded `Meta.drop` to traverse attached MetaLabel and MetaHeader data.
+  * Added `__delitem__` and `drop` to Instrument and Constellation classes.
   * Added options to customize `pysat_ndtesting` instrument with sample rate,
     shift in time.
   * Added orbit number to `pysat_ndtesting`.